requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
fastapi==0.110.2
fonttools==4.51.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kiwisolver==1.4.5
openai==1.23.2
packaging==24.0
pandas==2.2.1
pillow==10.3.0
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
